# proto/openapi/argosy.yaml
openapi: 3.0.3
info:
  title: Argosy API
  version: 0.1.0
  description: |
    Self-hosted media streaming server. This spec is the single source of truth
    for the HTTP contract; the Go server interface and the TypeScript client are
    generated from it (`make generate`), and CI fails on drift.
servers:
  - url: /
...
paths:
...
  /api/v1/scan:
    post:
      tags: [library]
      operationId: triggerScan
      summary: Trigger an immediate library re-scan ("rebuild the Manifest")
      responses:
        "202":
          description: A scan sweep was queued
        "409":
          description: A scan sweep is already running or queued
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/scan/status:
    get:
      tags: [library]
      operationId: getScanStatus
      summary: Current/last scan sweep status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Per-device token issued by registerDevice.
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ScanLibraryResult:
      type: object
      required: [libraryId, name, scanned, errors]
      properties:
        libraryId: { type: string, format: uuid }
        name: { type: string }
        scanned: { type: integer }
        errors: { type: integer }
        error: { type: string }
      additionalProperties: false
    ScanStatus:
      type: object
      required: [running, libraries]
      properties:
        running: { type: boolean }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
        libraries: { type: array, items: { $ref: "#/components/schemas/ScanLibraryResult" } }
...
security:
  - bearerAuth: []